Fix Memory allocation for large size above 2Gb

+++ src/aig/aig/aigCuts.c
@@ -64,7 +64,7 @@ Aig_ManCut_t * Aig_ManCutStart( Aig_Man_t * pMan, int nCutsMax, int nLeafMax, in
     // room for temporary truth tables
     if ( fTruth )
     {
-        p->puTemp[0] = ABC_ALLOC( unsigned, 4 * p->nTruthWords );
+        p->puTemp[0] = ABC_ALLOC( unsigned, 4 * (size_t)(p->nTruthWords ));
         p->puTemp[1] = p->puTemp[0] + p->nTruthWords;
         p->puTemp[2] = p->puTemp[1] + p->nTruthWords;
         p->puTemp[3] = p->puTemp[2] + p->nTruthWords;
